Done with the Booking Module

+++ src/main/java/com/oceanLink/service/users/UserService.java
@@ -1,6 +1,6 @@
 package com.oceanLink.service.users;
 
-import com.oceanLink.dto.UserDTO;
+import com.oceanLink.dto.user.UserDTO;
 import com.oceanLink.model.User;
 
 import java.util.List;
@@ -11,9 +11,9 @@ public interface UserService {
     User createUser(User users);
     List<UserDTO> findAllUser();
 
-    Optional<UserDTO> findUserById(int id);
+    Optional<UserDTO> findUserById(Long id);
 
-    Optional<UserDTO> updateUser(UserDTO userDTO,int ind);
+    Optional<UserDTO> updateUser(UserDTO userDTO,Long ind);
 
-    boolean deleteUser(int id);
+    boolean deleteUser(long id);
 }
